converted vumeter to tsx
